requirements.txt
annotated-types==0.6.0
anyio==3.7.1
certifi==2023.11.17
distro==1.8.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
idna==3.4
iniconfig==2.0.0
markdown-it-py==3.0.0
mdurl==0.1.2
openai==1.3.3
packaging==23.2
pluggy==1.3.0
pydantic==2.5.1
pydantic_core==2.14.3
Pygments==2.17.1
pyperclip==1.8.2
PyQt5==5.15.10
PyQt5-Qt5==5.15.11
PyQt5-sip==12.13.0
pytest==7.4.3
python-dotenv==1.0.0
rich==13.7.0
sniffio==1.3.0
tqdm==4.66.1
typing_extensions==4.8.0
